requirements.txt
fastapi
uvicorn
pymongo
pydantic
python-jose[cryptography]
bcrypt
motor